rows&columns: int to size_t

+++ tasks/seq/savchenko_m_min_matrix/include/ops_seq.hpp
@@ -8,20 +8,20 @@
 
 namespace savchenko_m_min_matrix_seq {
 
-std::vector<int> getRandomMatrix(int rows, int columns, int min, int max);
+std::vector<int> getRandomMatrix(size_t rows, size_t columns, int min, int max);
 
 class TestTaskSequential : public ppc::core::Task {
- public:
+public:
   explicit TestTaskSequential(std::shared_ptr<ppc::core::TaskData> taskData_) : Task(std::move(taskData_)) {}
   bool validation() override;
   bool pre_processing() override;
   bool run() override;
   bool post_processing() override;
 
- private:
+private:
   std::vector<int> matrix{};
   int res{};
-  int rows, columns;
+  size_t rows, columns;
 };
 
-}  // namespace savchenko_m_min_matrix_seq
+} // namespace savchenko_m_min_matrix_seq
