+++ tasks/seq/savchenko_m_min_matrix/src/ops_seq.cpp
@@ -6,14 +6,14 @@
 
 using namespace std::chrono_literals;
 
-std::vector<int> savchenko_m_min_matrix_seq::getRandomMatrix(int rows, int columns, int min, int max) {
+std::vector<int> savchenko_m_min_matrix_seq::getRandomMatrix(size_t rows, size_t columns, int min, int max) {
   std::random_device dev;
   std::mt19937 gen(dev());
 
   // Forming a random matrix
   std::vector<int> matrix(rows * columns);
-  for (int i = 0; i < rows; i++) {
-    for (int j = 0; j < columns; j++) {
+  for (size_t i = 0; i < rows; i++) {
+    for (size_t j = 0; j < columns; j++) {
       matrix[i * columns + j] = min + gen() % (max - min + 1);
     }
   }
@@ -34,10 +34,10 @@ bool savchenko_m_min_matrix_seq::TestTaskSequential::pre_processing() {
   rows = taskData->inputs_count[1];
   matrix = std::vector<int>(rows * columns);
 
-  for (int i = 0; i < rows; ++i) {
+  for (size_t i = 0; i < rows; ++i) {
     auto *temp = reinterpret_cast<int *>(taskData->inputs[i]);
 
-    for (int j = 0; j < columns; ++j) {
+    for (size_t j = 0; j < columns; ++j) {
       matrix[i * columns + j] = temp[j];
     }
   }
@@ -48,7 +48,7 @@ bool savchenko_m_min_matrix_seq::TestTaskSequential::pre_processing() {
 
 bool savchenko_m_min_matrix_seq::TestTaskSequential::run() {
   internal_order_test();
-  for (int i = 0; i < matrix.size(); i++) {
+  for (size_t i = 0; i < matrix.size(); i++) {
     if (matrix[i] < res) {
       res = matrix[i];
     }
